refactor: move refresh token logic to app
